\name{RcppMLPACK-package}
\alias{RcppMLPACK-package}
\alias{RcppMLPACK}
\docType{package}
\title{
  R integration with MLPACK using Rcpp
}
\description{
  The package eases the integration of MLPACK types with R.
}
\author{
  For RcppMLPACK: Qiang Kou
  
  For MLPACK: Ryan Curtin	
  
  Maintainer: Qiang Kou <qkou@umail.iu.edu>
}
\references{
  MLPACK project: \url{http://www.mlpack.org/}
}
\keyword{ package }
\keyword{ programming }
\keyword{ interface }


