\name{RcppMLPACK-package}
\alias{RcppMLPACK-package}
\alias{RcppMLPACK}
\docType{package}
\title{
  Rcpp/MLPACK bridge
}
\description{
  The package eases the integration of MLPACK types with Rcpp.
}
\author{
  For RcppMLPACK: Qiang Kou
	
  Maintainer: Qiang Kou <qkou@umail.iu.edu>

  For MLPACK: Ryan Curtin
}
\references{
  MLPACK project: \url{http://www.mlpack.org/}
}
\keyword{ package }
\keyword{ programming }
\keyword{ interface }


