\name{RcppMLPACK-package}
\alias{RcppMLPACK-package}
\alias{RcppMLPACK}
\docType{package}
\title{
  Rcpp Integration for MLPACK Library
}
\description{
  The package eases the integration of MLPACK types with R.
  MLPACK is an intuitive, fast, scalable C++ machine learning
   library, meant to be a machine learning analog to LAPACK.
}
\author{
  For RcppMLPACK: Qiang Kou
  
  For MLPACK: Ryan Curtin	
  
  Maintainer: Qiang Kou <qkou@umail.iu.edu>
}
\references{
  MLPACK project: \url{http://www.mlpack.org/}
}
\keyword{ package }
\keyword{ programming }
\keyword{ interface }


